+++ BooWho/client/src/Components/Auth/Register.js
@@ -1,28 +1,32 @@
 import React, { useState } from "react";
-import { useNavigate, Link } from "react-router-dom";
-import { login } from "../modules/authManager";
-
+import { useNavigate } from "react-router-dom";
+import { register } from "../../modules/authManager";
 import { LockClosedIcon } from "@heroicons/react/20/solid";
 
-export default function Login() {
-  let navigate = useNavigate();
+export default function Register() {
+  const navigate = useNavigate();
 
+  const [name, setName] = useState();
   const [email, setEmail] = useState();
   const [password, setPassword] = useState();
+  const [confirmPassword, setConfirmPassword] = useState();
 
   const handleNavigate = () => {
-    navigate("/register");
-  };
-
-  const handleHomeNavigate = () => {
     navigate("/boowho");
   };
 
-  const loginSubmit = (e) => {
+  const handleLoginNavigate = () => {
+    navigate("/login");
+  };
+
+  const registerClick = (e) => {
     e.preventDefault();
-    login(email, password)
-      .then(() => navigate("/boohome"))
-      .catch(() => alert("Login Failed, Boo"));
+    if (password && password !== confirmPassword) {
+      alert("Passwords don't match. Do better.");
+    } else {
+      const userProfile = { name, email };
+      register(userProfile, password).then(() => navigate("/"));
+    }
   };
 
   return (
@@ -57,11 +61,10 @@ export default function Login() {
         <div className="pt-1 float-right">
           <img
             className="h-auto w-auto object-center"
-            src={require("../assetts/friendghost.jpg")}
+            src={require("../assetts/ghostgirl.png")}
             alt=""
           />
         </div>
-
         <div className="flex min-h-full items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
           <div className="w-full max-w-md space-y-8">
             <div>
@@ -71,7 +74,7 @@ export default function Login() {
                 alt="BooWho"
               />
               <h2 className="mt-6 text-center text-4xl font-frijole tracking-tight text-white">
-                sign in to your account
+                register a new account
               </h2>
               <p className="mt-2 text-center text-sm text-white">
                 <div
@@ -81,11 +84,11 @@ export default function Login() {
                   <button
                     className="pt-10 text-lg font-cutive leading-6 text-yeller"
                     onClick={(event) => {
-                      handleNavigate(event);
+                      handleLoginNavigate(event);
                     }}
                   >
                     {" "}
-                    or register a new one
+                    or sign in if you already have one
                     <span aria-hidden="true">&rarr;</span>
                   </button>
                 </div>
@@ -94,6 +97,21 @@ export default function Login() {
             <form className="mt-8 space-y-6" action="#" method="POST">
               <input type="hidden" name="remember" defaultValue="true" />
               <div className="-space-y-px rounded-md shadow-sm">
+                <div className="pt-3">
+                  <label htmlFor="name" className="sr-only">
+                    Name
+                  </label>
+                  <input
+                    id="name"
+                    name="name"
+                    type="name"
+                    autoComplete="name"
+                    required
+                    className="relative block w-full appearance-none rounded-none rounded-t-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-500 focus:z-10 focus:border-indigo-500 focus:outline-none focus:ring-indigo-500 sm:text-sm"
+                    placeholder="Name"
+                    onChange={(e) => setName(e.target.value)}
+                  />
+                </div>
                 <div>
                   <label htmlFor="email-address" className="sr-only">
                     Email address
@@ -104,12 +122,12 @@ export default function Login() {
                     type="email"
                     autoComplete="email"
                     required
-                    className="relative block w-full appearance-none rounded-none rounded-t-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-500 focus:z-10 focus:border-indigo-500 focus:outline-none focus:ring-indigo-500 sm:text-sm"
+                    className="relative block w-full appearance-none rounded-none rounded-b-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-500 focus:z-10 focus:border-indigo-500 focus:outline-none focus:ring-indigo-500 sm:text-sm"
                     placeholder="Email address"
                     onChange={(e) => setEmail(e.target.value)}
                   />
                 </div>
-                <div>
+                <div className="pt-3">
                   <label htmlFor="password" className="sr-only">
                     Password
                   </label>
@@ -117,20 +135,33 @@ export default function Login() {
                     id="password"
                     name="password"
                     type="password"
-                    autoComplete="current-password"
                     required
-                    className="relative block w-full appearance-none rounded-none rounded-b-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-500 focus:z-10 focus:border-indigo-500 focus:outline-none focus:ring-indigo-500 sm:text-sm"
+                    className="relative block w-full appearance-none rounded-none rounded-t-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-500 focus:z-10 focus:border-indigo-500 focus:outline-none focus:ring-indigo-500 sm:text-sm"
                     placeholder="Password"
                     onChange={(e) => setPassword(e.target.value)}
                   />
                 </div>
+                <div>
+                  <label htmlFor="confirmPassword" className="sr-only">
+                    Password
+                  </label>
+                  <input
+                    id="confirmPassword"
+                    name="confirmPassword"
+                    type="password"
+                    required
+                    className="relative block w-full appearance-none rounded-none rounded-b-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-500 focus:z-10 focus:border-indigo-500 focus:outline-none focus:ring-indigo-500 sm:text-sm"
+                    placeholder="Confirm Password"
+                    onChange={(e) => setConfirmPassword(e.target.value)}
+                  />
+                </div>
               </div>
 
-              <div className="pb-40">
+              <div>
                 <button
                   type="submit"
                   className="group relative flex w-full justify-center rounded-md border border-transparent bg-indigo-600 py-2 px-4 text-sm font-medium text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                  onClick={loginSubmit}
+                  onClick={registerClick}
                 >
                   <span className="absolute inset-y-0 left-0 flex items-center pl-3">
                     <LockClosedIcon
@@ -138,26 +169,26 @@ export default function Login() {
                       aria-hidden="true"
                     />
                   </span>
-                  Sign in
+                  Register
                 </button>
-                <p className="mt-2 text-center text-sm text-white">
-                  <div
-                    href="#!"
-                    className="text-md font-medium text-yeller hover:text-white"
-                  >
-                    <button
-                      className="pt-10 pb-10 text-lg font-cutive leading-6 text-yeller"
-                      onClick={(event) => {
-                        handleHomeNavigate(event);
-                      }}
-                    >
-                      {" "}
-                      take me back home
-                      <span aria-hidden="true">&rarr;</span>
-                    </button>
-                  </div>
-                </p>
               </div>
+              <p className="mt-2 text-center text-sm text-white">
+                <div
+                  href="#!"
+                  className="text-md font-medium text-yeller hover:text-white"
+                >
+                  <button
+                    className="pt-10 pb-10 text-lg font-cutive leading-6 text-yeller"
+                    onClick={(event) => {
+                      handleNavigate(event);
+                    }}
+                  >
+                    {" "}
+                    take me back home
+                    <span aria-hidden="true">&rarr;</span>
+                  </button>
+                </div>
+              </p>
             </form>
           </div>
         </div>
