+++ BooWho/client/src/Components/House/HouseAddForm.js
@@ -1,0 +1,30 @@
+import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
+
+export default function HouseForm() {
+  let navigate = useNavigate();
+
+  const [email, setEmail] = useState();
+  const [password, setPassword] = useState();
+
+  const handleNavigate = () => {
+    navigate("/register");
+  };
+
+  const handleHomeNavigate = () => {
+    navigate("/boowho");
+  };
+
+  const loginSubmit = (e) => {
+    e.preventDefault();
+    login(email, password)
+      .then(() => navigate("/boohome"))
+      .catch(() => alert("Login Failed, Boo"));
+  };
+
+  return (
+    <>
+      <div className="isolate bg-spooky"></div>
+    </>
+  );
+}
