+++ BooWho/client/src/Components/House/HouseList.js
@@ -1,71 +1,102 @@
-const products = [
-  {
-    id: 1,
-    name: "Earthen Bottle",
-    href: "#",
-    price: "$48",
-    imageSrc:
-      "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-01.jpg",
-    imageAlt:
-      "Tall slender porcelain bottle with natural clay textured body and cork stopper.",
-  },
-  {
-    id: 2,
-    name: "Nomad Tumbler",
-    href: "#",
-    price: "$35",
-    imageSrc:
-      "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-02.jpg",
-    imageAlt:
-      "Olive drab green insulated bottle with flared screw lid and flat top.",
-  },
-  {
-    id: 3,
-    name: "Focus Paper Refill",
-    href: "#",
-    price: "$89",
-    imageSrc:
-      "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-03.jpg",
-    imageAlt:
-      "Person using a pen to cross a task off a productivity paper card.",
-  },
-  {
-    id: 4,
-    name: "Machined Mechanical Pencil",
-    href: "#",
-    price: "$35",
-    imageSrc:
-      "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-04.jpg",
-    imageAlt:
-      "Hand holding black machined steel mechanical pencil with brass tip and top.",
-  },
-  // More products...
-];
+import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
 
 export default function HouseList() {
-  return (
-    <div className="bg-white">
-      <div className="mx-auto max-w-2xl py-16 px-4 sm:py-24 sm:px-6 lg:max-w-7xl lg:px-8">
-        <h2 className="sr-only">Products</h2>
+  let navigate = useNavigate();
 
-        <div className="grid grid-cols-1 gap-y-10 gap-x-6 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 xl:gap-x-8">
-          {products.map((product) => (
-            <a key={product.id} href={product.href} className="group">
-              <div className="aspect-w-1 aspect-h-1 w-full overflow-hidden rounded-lg bg-gray-200 xl:aspect-w-7 xl:aspect-h-8">
-                <img
-                  src={product.imageSrc}
-                  alt={product.imageAlt}
-                  className="h-full w-full object-cover object-center group-hover:opacity-75"
-                />
-              </div>
-              <h3 className="mt-4 text-sm text-gray-700">{product.name}</h3>
-              <p className="mt-1 text-lg font-medium text-gray-900">
-                {product.price}
-              </p>
-            </a>
-          ))}
-        </div>
-      </div>
-    </div>
+  const [email, setEmail] = useState();
+  const [password, setPassword] = useState();
+
+  const handleNavigate = () => {
+    navigate("/register");
+  };
+
+  const handleHomeNavigate = () => {
+    navigate("/boowho");
+  };
+
+  const loginSubmit = (e) => {
+    e.preventDefault();
+    login(email, password)
+      .then(() => navigate("/boohome"))
+      .catch(() => alert("Login Failed, Boo"));
+  };
+
+  return (
+    <>
+      <div className="isolate bg-spooky"></div>
+    </>
   );
 }
+
+// const products = [
+//   {
+//     id: 1,
+//     name: "Earthen Bottle",
+//     href: "#",
+//     price: "$48",
+//     imageSrc:
+//       "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-01.jpg",
+//     imageAlt:
+//       "Tall slender porcelain bottle with natural clay textured body and cork stopper.",
+//   },
+//   {
+//     id: 2,
+//     name: "Nomad Tumbler",
+//     href: "#",
+//     price: "$35",
+//     imageSrc:
+//       "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-02.jpg",
+//     imageAlt:
+//       "Olive drab green insulated bottle with flared screw lid and flat top.",
+//   },
+//   {
+//     id: 3,
+//     name: "Focus Paper Refill",
+//     href: "#",
+//     price: "$89",
+//     imageSrc:
+//       "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-03.jpg",
+//     imageAlt:
+//       "Person using a pen to cross a task off a productivity paper card.",
+//   },
+//   {
+//     id: 4,
+//     name: "Machined Mechanical Pencil",
+//     href: "#",
+//     price: "$35",
+//     imageSrc:
+//       "https://tailwindui.com/img/ecommerce-images/category-page-04-image-card-04.jpg",
+//     imageAlt:
+//       "Hand holding black machined steel mechanical pencil with brass tip and top.",
+//   },
+//   // More products...
+// ];
+
+// export default function HouseList() {
+//   return (
+//     <div className="bg-white">
+//       <div className="mx-auto max-w-2xl py-16 px-4 sm:py-24 sm:px-6 lg:max-w-7xl lg:px-8">
+//         <h2 className="sr-only">Products</h2>
+
+//         <div className="grid grid-cols-1 gap-y-10 gap-x-6 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 xl:gap-x-8">
+//           {products.map((product) => (
+//             <a key={product.id} href={product.href} className="group">
+//               <div className="aspect-w-1 aspect-h-1 w-full overflow-hidden rounded-lg bg-gray-200 xl:aspect-w-7 xl:aspect-h-8">
+//                 <img
+//                   src={product.imageSrc}
+//                   alt={product.imageAlt}
+//                   className="h-full w-full object-cover object-center group-hover:opacity-75"
+//                 />
+//               </div>
+//               <h3 className="mt-4 text-sm text-gray-700">{product.name}</h3>
+//               <p className="mt-1 text-lg font-medium text-gray-900">
+//                 {product.price}
+//               </p>
+//             </a>
+//           ))}
+//         </div>
+//       </div>
+//     </div>
+//   );
+// }
